FPDS Get New

A convenience function to compare a recently captured DOGE object with data previously downloaded by FPDS. This will run requests to FPDS based only on new DOGE rows captured

Usage

fpds_get_new(
  doge_data = NULL,
  fpds_data = NULL,
  verbose = FALSE,
  return = "combined"
)

Arguments

doge_data

A data object created by doge_get_data()

fpds_data

A data object created by fpds_get_data()

verbose

Logical, whether to return messages.

return

Whether to return a merged file with new and old data, or simply the new data.

Value

A named list comrpised of data.frames.

Examples

if (FALSE) { # \dontrun{
fpds_get_new(doge_data,fpds_data)
} # }